Convert all references to EntityQueryResult to QueryResult
Also integrated the special "FullResult" constructor, and tests from
EQR
