+++ lib/models/src/main/java/org/sagebionetworks/repo/model/QueryResults.java
@@ -1,10 +1,7 @@
 package org.sagebionetworks.repo.model;
 
-import java.io.Serializable;
 import java.util.LinkedHashMap;
 import java.util.List;
-
-import javax.xml.bind.annotation.XmlRootElement;
 
 import org.sagebionetworks.schema.ObjectSchema;
 import org.sagebionetworks.schema.TYPE;
@@ -60,6 +57,30 @@ public class QueryResults<T> {
 			long totalNumberOfResults) {
 		this.results = results;
 		this.totalNumberOfResults = totalNumberOfResults;
+	}
+
+	/**
+	 * Given a full list of results and pagination parameters, creates a EntityQueryResults the subList matching the pagination parameters.
+	 * @param fullResults - Should be the full list, not just one page.  The fullResults.size() will be used for the totalNumberOfResults.
+	 * @param limit - Sets the page size.
+	 * @param offest - Sets the start of the page.
+	 */
+	public QueryResults(List<T> fullResults, int limit, int offest) {
+		super();
+		if(fullResults == null) throw new IllegalArgumentException("FullResults cannot be null");
+		if(offest < 0) throw new IllegalArgumentException("Offset cannot be less than zero");
+		if(limit < 0) throw new IllegalArgumentException("Limit cannot be less than zero");
+		// Calculate the indices
+		// The start is inclusive in List.subList();
+		int startIndex = offest;
+		// The end is exclusive in List.SubList()
+		int endIndex = offest + limit;
+		// Note, if limit is Integer.MAX + then offest + limit will be negative.
+		if(endIndex > fullResults.size() || endIndex < 0){
+			endIndex = fullResults.size();
+		}
+		this.results = fullResults.subList(startIndex, endIndex);
+		this.totalNumberOfResults = fullResults.size();
 	}
 
 	/**
